rename: api-domains-fishing-licence -> api-domains-fishing-license
