+++ libs/api/domains/fishing-license/.eslintrc.json
@@ -1,5 +1,5 @@
 {
-  "extends": ["../../../../.eslintrc.json"],
+  "extends": "../../../../.eslintrc.json",
   "ignorePatterns": ["!**/*"],
   "rules": {}
 }
