+++ libs/api/domains/fishing-license/src/graphql/models/deprivation.model.ts
