+++ libs/api/domains/fishing-license/src/graphql/models/fishing-license-info.model.ts
