+++ libs/api/domains/fishing-license/src/graphql/models/seaworthiness.model.ts
