+++ libs/api/domains/fishing-license/src/graphql/models/ship.model.ts
