+++ libs/api/domains/fishing-license/tsconfig.lib.json
