+++ libs/api/domains/fishing-license/tsconfig.spec.json
