+++ nx.json
@@ -121,7 +121,7 @@
     "api-domains-finance": {
       "tags": []
     },
-    "api-domains-fishing-licence": {
+    "api-domains-fishing-license": {
       "tags": []
     },
     "api-domains-health-insurance": {
